SMBIOS update
We can now use a double -1 value for SMBProperties[x].start and
SMBProperties[x].stop to signal that there isn;'t anything we like to
add/override, but the complete structure from the factory table is
copied.

The memory overrides are expanded. Notably the error handle feature.

We now add a type 11 (kSMBTypeOemPlatformFeature) table so that
AppleSMBIOS.kext can use this data and add IOService:/platform-feature.